requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
certifi=2023.5.7=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=39.0.1=py310h21b164f_2
cuda-cccl=12.2.53=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
filelock=3.9.0=py310haa95532_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
idna=3.4=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46319
jinja2=3.1.2=py310haa95532_0
jpeg=9e=h2bbff1b_1
lerc=3.0=hd77b12b_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.53=0
libcurand-dev=10.3.3.53=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lightning-utilities=0.9.0=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py310h2bbff1b_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.6=py310h4ed8f06_1
mkl_random=1.2.2=py310h4ed8f06_1
mpmath=1.2.1=py310haa95532_0
networkx=2.8.4=py310haa95532_1
numpy=1.25.0=py310h055cbcc_0
numpy-base=1.25.0=py310h65a83cf_0
openssl=3.0.9=h2bbff1b_0
packaging=23.1=pypi_0
pillow=9.4.0=py310hd77b12b_0
pip=23.1.2=py310haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py310haa95532_0
pysocks=1.7.1=py310haa95532_0
python=3.10.11=he1021f5_3
pytorch=2.0.1=py3.10_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-mutex=1.0=cuda
requests=2.29.0=py310haa95532_0
setuptools=67.8.0=py310haa95532_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.11.1=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.12=h2bbff1b_0
torchaudio=2.0.2=pypi_0
torchmetrics=1.0.0=pypi_0
torchvision=0.15.2=pypi_0
typing_extensions=4.6.3=py310haa95532_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.2=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
